Re-write of Database Manager
Change the format of the Enumerator DownloadState (no more queued).
Gave each of the properties of MediaInfo a default value.
Updated the the getEpisode method of MediaInfo due to float related
errors.
Added local final path for media as well as remote for WDMyCloud.
Also clean u a couple of TODOs:
- Finish implementing this (checkIfNetworkLocationAvailable) method
- Convert all System.out.print statements to log.info
- Convert failed and downloading items in database back te pending
- Clean up the writing to the database and the handling of MediaInfo
Objects