Fixed the "write download to file" bug, the "can't resolve hash for download" bug, the "no slashes in path for database entries" bug
The write file to download bug was occuring for media which did not have
a file type in the url.
The "cant resolve hash for download" bug was occurring becuase the media
page was loading up without redirect to the "Proceed to video" page,
which media manager wasn't expecting.
The "no slashes in path for database entries" bug was occurring because
the path entries were being filtered for invalid characters which was
filtering the slashes out.
The "mis-parsing JSON strings to the Web App" bug was occurring because
quotes in movie titles were not being paired with an escape character.
Also added a unit test to test future urls.